move stickyMessage feature to features folder
