+++ src/Bot.ts
@@ -4,7 +4,7 @@ import commands from './commands/index'
 import { Environment } from './config'
 import eventPlugins from './events/index'
 import featurePlugins from './features'
-import { initStickyMessage } from './features/stickyMessage/index'
+import { initStickyMessage } from './features/stickyMessage/stickyMessages'
 import { BotContext } from './types/BotContext'
 import { CommandHandlerConfig } from './types/CommandHandlerConfig'
 import { Plugin } from './types/Plugin'
