+++ src/features/index.ts
@@ -1,5 +1,7 @@
 import { Plugin } from '@/types/Plugin'
 
 import nameChecker from './nameChecker'
+import preventEmojiSpam from './preventEmojiSpam'
+import stickyMessage from './stickyMessage'
 
-export default [nameChecker] as Plugin[]
+export default [nameChecker, preventEmojiSpam, stickyMessage] as Plugin[]
